server/requirements.txt
openenv-core[core]>=0.2.3
fastapi>=0.115.0
uvicorn>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0
